refactor: refacto bots service
Move bots files to bots folder, use simulation name when grSim is not needed. Add action schema for
order and placement in their own folder
