+++ src/bots/grSim/compiled/compiledProtobufGrSimPacket.d.ts
