+++ src/bots/grSim/compiled/compiledProtobufGrSimPacket.js
