+++ src/bots/grSim/compiled/compiledProtobufGrSimRobotStatus.d.ts
