+++ src/bots/grSim/compiled/compiledProtobufGrSimRobotStatus.js
