+++ src/bots/grSim/index.ts
