+++ src/bots/grSim/proto/messages_robocup_ssl_robot_status.proto
